+++ menu/menu_internals.h
@@ -6,7 +6,7 @@
 /*   By: frapp <fabi@student.42.fr>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/01 14:21:36 by frapp             #+#    #+#             */
-/*   Updated: 2024/05/26 03:14:52 by frapp            ###   ########.fr       */
+/*   Updated: 2024/06/16 06:45:13 by frapp            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@ void	close_menu(t_main *main_data, t_menu *menu);
 void	open_menu(t_main *main_data);
 void	handle_open_menu(t_main *main_data, t_menu *menu);
 void	init_entry_widget(t_main *main_data, t_entry_widget *field, char *title, int depth);
-void	fill_slider(t_entry_widget *slider, float filling);
+void	fill_slider(t_entry_widget *slider, double filling);
 
 #endif
